AppInboundProtocol: relocate to application directory
