+++ xbmc/input/InputManager.cpp
@@ -8,7 +8,6 @@
 
 #include "InputManager.h"
 
-#include "AppInboundProtocol.h"
 #include "Application.h"
 #include "ButtonTranslator.h"
 #include "CustomControllerTranslator.h"
@@ -19,6 +18,7 @@
 #include "TouchTranslator.h"
 #include "Util.h"
 #include "XBMC_vkeys.h"
+#include "application/AppInboundProtocol.h"
 #include "application/AppParamParser.h"
 #include "guilib/GUIAudioManager.h"
 #include "guilib/GUIComponent.h"
