+++ xbmc/input/touch/generic/GenericTouchActionHandler.cpp
@@ -8,8 +8,8 @@
 
 #include "GenericTouchActionHandler.h"
 
-#include "AppInboundProtocol.h"
 #include "ServiceBroker.h"
+#include "application/AppInboundProtocol.h"
 #include "guilib/GUIComponent.h"
 #include "guilib/GUIWindowManager.h"
 #include "input/Key.h"
