+++ xbmc/platform/darwin/ios/IOSEAGLView.mm
@@ -8,13 +8,13 @@
 
 #import "IOSEAGLView.h"
 
-#include "AppInboundProtocol.h"
 #include "Application.h"
 #import "IOSScreenManager.h"
 #include "ServiceBroker.h"
 #include "Util.h"
 #import "XBMCController.h"
 #include "application/AppEnvironment.h"
+#include "application/AppInboundProtocol.h"
 #include "application/AppParams.h"
 #include "messaging/ApplicationMessenger.h"
 #include "settings/AdvancedSettings.h"
