+++ xbmc/platform/linux/input/LIRC.cpp
@@ -8,8 +8,8 @@
 
 #include "LIRC.h"
 
-#include "AppInboundProtocol.h"
 #include "ServiceBroker.h"
+#include "application/AppInboundProtocol.h"
 #include "profiles/ProfileManager.h"
 #include "settings/AdvancedSettings.h"
 #include "settings/Settings.h"
