+++ xbmc/platform/linux/input/LibInputPointer.cpp
@@ -8,8 +8,8 @@
 
 #include "LibInputPointer.h"
 
-#include "AppInboundProtocol.h"
 #include "ServiceBroker.h"
+#include "application/AppInboundProtocol.h"
 #include "input/mouse/MouseStat.h"
 #include "utils/log.h"
 #include "windowing/GraphicContext.h"
