+++ xbmc/windowing/X11/WinEventsX11.cpp
@@ -8,9 +8,9 @@
 
 #include "WinEventsX11.h"
 
-#include "AppInboundProtocol.h"
 #include "Application.h"
 #include "ServiceBroker.h"
+#include "application/AppInboundProtocol.h"
 #include "cores/AudioEngine/Interfaces/AE.h"
 #include "guilib/GUIComponent.h"
 #include "guilib/GUIWindowManager.h"
