+++ xbmc/windowing/android/WinEventsAndroid.cpp
@@ -8,8 +8,8 @@
 
 #include "WinEventsAndroid.h"
 
-#include "AppInboundProtocol.h"
 #include "ServiceBroker.h"
+#include "application/AppInboundProtocol.h"
 #include "guilib/GUIComponent.h"
 #include "guilib/GUIWindowManager.h"
 #include "input/InputManager.h"
