+++ xbmc/windowing/osx/SDL/WinEventsSDL.cpp
@@ -8,10 +8,10 @@
 
 #include "WinEventsSDL.h"
 
-#include "AppInboundProtocol.h"
 #include "Application.h"
 #include "GUIUserMessages.h"
 #include "ServiceBroker.h"
+#include "application/AppInboundProtocol.h"
 #include "guilib/GUIComponent.h"
 #include "guilib/GUIWindowManager.h"
 #include "input/InputManager.h"
