+++ xbmc/windowing/wayland/WinEventsWayland.cpp
@@ -8,8 +8,8 @@
 
 #include "WinEventsWayland.h"
 
-#include "AppInboundProtocol.h"
 #include "ServiceBroker.h"
+#include "application/AppInboundProtocol.h"
 #include "threads/CriticalSection.h"
 #include "threads/Thread.h"
 #include "utils/log.h"
